Fixing user profile workflow
I encountered problems in which the Devise controllers were not
rendering the proper layout. This was fixed by switching the devise
controllers to include the with_themed_layout.

Closes ndlib/planning#80
