+++ app/controllers/concerns/curate_controller.rb
@@ -20,26 +20,10 @@ module CurateController
 
   included do
     include BreadcrumbsOnRails::ActionController
+    include Curate::ThemedLayoutController
     add_breadcrumb "Dashboard", :dashboard_index_path
 
-    class_attribute :theme
-    self.theme = 'curate_nd'
     rescue_from StandardError, with: :exception_handler
-
-    helper_method :theme
-    helper_method :show_action_bar?
-    helper_method :show_site_search?
-  end
-
-
-  module ClassMethods
-    def with_themed_layout(view_name = nil)
-      if view_name
-        layout("#{theme}/#{view_name}")
-      else
-        layout(theme)
-      end
-    end
   end
 
 
@@ -75,19 +59,6 @@ module CurateController
     render "/errors/#{wrapper.status_code}", status: wrapper.status_code, layout: !request.xhr?
   end
   protected :render_response_for_error
-
-
-  # layout 'hydra-head'
-
-  # protect_from_forgery
-
-  def show_action_bar?
-    true
-  end
-
-  def show_site_search?
-    true
-  end
 
   protected
