+++ app/controllers/sessions_controller.rb
@@ -1,0 +1,4 @@
+class SessionsController < Devise::SessionsController
+  include Curate::ThemedLayoutController
+  with_themed_layout '1_column'
+end
